backend/resume_scanner/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
PyMuPDF
pandas